# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wheel Strategy Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Settings {
  tickers: string[];
  /** @minimum 1 */
  minDte: number;
  /** @minimum 1 */
  maxDte: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  targetDelta: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  minDelta: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  maxDelta: number;
  /** @minimum 0 */
  minOpenInterest: number;
  /** @minimum 0 */
  minBid: number;
  /** @minimum 0 */
  minUnderlyingPrice: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  riskFreeRate: number;
  /**
   * @minimum 1
   * @maximum 200
   */
  topN: number;
  /**
   * @minimum 1
   * @maximum 1440
   */
  cacheTtlMinutes: number;
}

export interface SettingsInput {
  tickers: string[];
  /** @minimum 1 */
  minDte: number;
  /** @minimum 1 */
  maxDte: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  targetDelta: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  minDelta: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  maxDelta: number;
  /** @minimum 0 */
  minOpenInterest: number;
  /** @minimum 0 */
  minBid: number;
  /** @minimum 0 */
  minUnderlyingPrice: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  riskFreeRate: number;
  /**
   * @minimum 1
   * @maximum 200
   */
  topN: number;
  /**
   * @minimum 1
   * @maximum 1440
   */
  cacheTtlMinutes: number;
}

/**
 * Optional one-shot overrides for this scan (does not persist)
 */
export interface ScanInput {
  tickers?: string[];
  /** @minimum 1 */
  minDte?: number;
  /** @minimum 1 */
  maxDte?: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  targetDelta?: number;
  forceRefresh?: boolean;
}

export interface Candidate {
  ticker: string;
  spot: number;
  /** ISO date YYYY-MM-DD */
  expiry: string;
  dte: number;
  strike: number;
  delta: number;
  bid: number;
  /** @nullable */
  ask?: number | null;
  /** Implied volatility, 0..1 */
  iv: number;
  openInterest: number;
  /** @nullable */
  volume?: number | null;
  premiumPerContract: number;
  collateralPerContract: number;
  staticReturnPct: number;
  annualizedPct: number;
  breakeven: number;
  pctOtm: number;
  /**
   * 0..1, proxy from realized vol
   * @nullable
   */
  ivRank?: number | null;
  /** @nullable */
  hv30?: number | null;
  /** @nullable */
  earningsDate?: string | null;
  earningsInWindow: boolean;
}

export interface ScanError {
  ticker: string;
  reason: string;
}

export interface ScanResult {
  /**
   * ISO timestamp of when this scan ran
   * @nullable
   */
  scannedAt: string | null;
  candidates: Candidate[];
  errors: ScanError[];
  tickersScanned: number;
  tickersWithCandidate: number;
  cached: boolean;
}

export interface IvRankBucket {
  label: string;
  count: number;
}

export interface ScanSummary {
  /** @nullable */
  scannedAt: string | null;
  candidateCount: number;
  earningsFlaggedCount: number;
  avgAnnualizedPct: number;
  medianAnnualizedPct: number;
  maxAnnualizedPct: number;
  totalCollateral: number;
  totalPremium: number;
  ivRankBuckets: IvRankBucket[];
  /** Top 5 candidates by annualized yield */
  topTickers: Candidate[];
}

export interface Quote {
  ticker: string;
  spot: number;
  asOf: string;
  /** @nullable */
  earningsDate?: string | null;
  earningsInWindow: boolean;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  dayChange?: number | null;
  /** @nullable */
  dayChangePct?: number | null;
}

export interface ChainExpirations {
  ticker: string;
  spot: number;
  expirations: string[];
}

export interface ChainRow {
  strike: number;
  bid: number;
  ask: number;
  lastPrice: number;
  iv: number;
  delta: number;
  gamma: number;
  theta: number;
  vega: number;
  openInterest: number;
  volume: number;
  inTheMoney: boolean;
}

export interface Chain {
  ticker: string;
  spot: number;
  expiry: string;
  dte: number;
  puts: ChainRow[];
  calls: ChainRow[];
}