# lib/api-zod/src/generated/types/candidate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wheel Strategy Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Candidate {
  ticker: string;
  spot: number;
  /** ISO date YYYY-MM-DD */
  expiry: string;
  dte: number;
  strike: number;
  delta: number;
  bid: number;
  /** @nullable */
  ask?: number | null;
  /** Implied volatility, 0..1 */
  iv: number;
  openInterest: number;
  /** @nullable */
  volume?: number | null;
  premiumPerContract: number;
  collateralPerContract: number;
  staticReturnPct: number;
  annualizedPct: number;
  breakeven: number;
  pctOtm: number;
  /**
   * 0..1, proxy from realized vol
   * @nullable
   */
  ivRank?: number | null;
  /** @nullable */
  hv30?: number | null;
  /** @nullable */
  earningsDate?: string | null;
  earningsInWindow: boolean;
}