# lib/api-zod/src/generated/types/chain.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wheel Strategy Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ChainRow } from "./chainRow";

export interface Chain {
  ticker: string;
  spot: number;
  expiry: string;
  dte: number;
  puts: ChainRow[];
  calls: ChainRow[];
}