# lib/api-zod/src/generated/types/chainRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wheel Strategy Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ChainRow {
  strike: number;
  bid: number;
  ask: number;
  lastPrice: number;
  iv: number;
  delta: number;
  gamma: number;
  theta: number;
  vega: number;
  openInterest: number;
  volume: number;
  inTheMoney: boolean;
}