# lib/api-zod/src/generated/types/quote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wheel Strategy Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Quote {
  ticker: string;
  spot: number;
  asOf: string;
  /** @nullable */
  earningsDate?: string | null;
  earningsInWindow: boolean;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  dayChange?: number | null;
  /** @nullable */
  dayChangePct?: number | null;
}