# lib/api-zod/src/generated/types/scanError.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wheel Strategy Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ScanError {
  ticker: string;
  reason: string;
}