# lib/api-zod/src/generated/types/scanInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wheel Strategy Dashboard API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Optional one-shot overrides for this scan (does not persist)
 */
export interface ScanInput {
  tickers?: string[];
  /** @minimum 1 */
  minDte?: number;
  /** @minimum 1 */
  maxDte?: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  targetDelta?: number;
  forceRefresh?: boolean;
}